Merged PR 70: Create a biceps for resources
This PR shows the provisions that wanted cloud resources in our HopeLearnBridge project using bicep , its  a declarative  Infrastructure as a code language. Bicep helps us define , configure and deploy our resource using code as follows:

1 - ResourceGroup.bicep
Resource group act as a container for application's resources named HopeLearnRGroup in the eastus location

2- AppServicePlan.bicep
It defines the OS of the resources by kind parameter and it determines the features and their cost by sku parameter.
There are two function apps within the plan
they are configured with the plan through serverFrameId ,verified https communication by setting httpsonly true. In addition to defining technology stack versions using siteconfig parameters

3- Cosmosdb.bicep

The account named using unique function  to generate a unique name for the account (I saw that recommended over the internet). The other name is for the database within that account
4- main.bicep
It contains two modules one for AppServicePlan and the other for Cosmosdb

1- Bicep compilation :  Bicep files are compiled into ARM templates
2- Resource creation : Azure resource manager parses the ARM template and create the specified resources in the Azure environment
3- Resource configuration : ARM applies the desired configurations to the resources such as setting the sku , properties .. etc
4-  Deployment completion : Once all resources are created and configured ; deployment is completed
**az deployment sub create --location eastus --template-file resourceGroup.bicep**

 **az deployment group create --resource-group rg-hlp-dev-eus-001 --template-file main.bicep -p main.parameters.development.json**

Related work items: #137