Merged PR 110: course-creation-role-based-access
I implemented the functionality to ensure that when a user creates a course, only users with the role of "Teacher" are allowed to create courses. This restriction ensures that only teachers can create courses in the system. Additionally, when a user requests to view courses, the displayed courses are based on their role. If the user is a "Student," they will see all available courses. However, if the user is a "Teacher," they will only see the courses they have created.

_____________________________________

I tested all the endpoints, and everything is working correctly. The role-based logic is functioning as expected: teachers can create courses, and course retrieval is based on the user's role (students see all courses, teachers see only the courses they created).

___________________________________

![image.png](https://dev.azure.com/ASAL-UniversityTraining/228d1bc8-1a6f-4ac8-b695-5c5b9c4f261f/_apis/git/repositories/c3f3f6e6-46dc-448e-99fd-6e88db5435a5/pullRequests/110/attachments/image.png)

__________________________

![image (2).png](https://dev.azure.com/ASAL-UniversityTraining/228d1bc8-1a6f-4ac8-b695-5c5b9c4f261f/_apis/git/repositories/c3f3f6e6-46dc-448e-99fd-6e88db5435a5/pullRequests/110/attachments/image%20%282%29.png)
_____________________________

![image (3).png](https://dev.azure.com/ASAL-UniversityTraining/228d1bc8-1a6f-4ac8-b695-5c5b9c4f261f/_apis/git/repositories/c3f3f6e6-46dc-448e-99fd-6e88db5435a5/pullRequests/110/attachments/image%20%283%29.png)

Related work items: #184